04_gtw/container_creation/requirements.txt
#pip list
paho-mqtt>=1.6.1
psycopg2>=2.9.3
#psycopg2-binary>=2.9.2
